Move MenuBarBatteryDot into the test-compiled spacing helpers.
